+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - refactor(services): make sure services do not do anything, including importing their `setup.py` which can potentially run code in its module scope, by moving `register_reducer(reducer)` write after reducer import, before `init_service` is imported
 - chore(scripts): add `--index` for `device:` poe scripts, allowing deployment and tests on multiple devices without the need to constantly change `ubo-development-pod`
 - chore: housekeeping, update dependencies, remove unused code, etc
+- feat(infrared): add infrared service with settings menu to enable/disable propagating keypad actions as ir commands and receiving ir commands and translating them to keypad actions - closes #160
 
 ## Version 1.3.0